% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correction.R
\name{batch_correction}
\alias{batch_correction}
\title{Batch Correction}
\usage{
batch_correction(
  mat = NULL,
  meta = NULL,
  method,
  batch.1 = NULL,
  batch.2 = NULL,
  log2_tranformed = TRUE,
  treatment = NULL,
  housekeeping = NULL,
  k = 2,
  drop = 0,
  center = FALSE,
  round = FALSE,
  tolerance = 1e-08,
  par.prior = TRUE
)
}
\arguments{
\item{mat}{Numeric matrix with features as rownames and sample names as the column names}

\item{meta}{Data frame of sample data with the first column being sample names that match the column names of the matrix}

\item{method}{A character vector of batch correction methods}

\item{batch.1}{Column name from the meta file of the column that will be used for batch one information}

\item{batch.2}{Column name from the meta file of the column that will be used for batch two information}

\item{log2_tranformed}{logical whether the data is alrady transformed}

\item{treatment}{Column name from the meta file of the column that will be used for treatment information}

\item{housekeeping}{Name of housekeeping gene set or character vector of housekeeping genes}

\item{k}{Used in the RUVg method, the number of factors of unwanted variation to be estimated from the data}

\item{drop}{Used in the RUVg method, the number of singular values to drop in the estimation of the factors of unwanted variation. This number is usually zero, but might be set to one if the first singular value captures the effect of interest. It must be less than k}

\item{center}{Used in the RUVg method, if TRUE, the counts are centered, for each gene, to have mean zero across samples. This is important to ensure that the first singular value does not capture the average gene expression}

\item{round}{Used in the RUVg method, if TRUE, the normalized measures are rounded to form pseudo-counts}

\item{tolerance}{Used in the RUVg method, tolerance in the selection of the number of positive singular values, i.e., a singular value must be larger than tolerance to be considered positive}

\item{par.prior}{Used in the ComBat method, TRUE indicates parametric adjustments will be used, FALSE indicates non-parametric adjustments will be used}
}
\value{
List object of length of method
}
\description{
Batch Correction
}
\examples{
set.seed(101)
}
