% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch_FLEX.R
\name{Batch_FLEX}
\alias{Batch_FLEX}
\title{BatchFLEX}
\usage{
Batch_FLEX(
  Batch_FLEX_function = c("batch_correct", "batch_evaluate"),
  num_samples = 1000,
  num_genes = 20000,
  num_control = 18000,
  num_treatments = 2,
  treatment_effect_multiplier = c(0, 1),
  treatment_effect = 5,
  treatment_sd = 1,
  additional_batch = 1,
  batch_proportion = 0.5,
  batch_effect_multiplier = c(1),
  batch_effect = 5,
  batch_effect_sd = 1,
  batch_sample_sd_mean = 0,
  batch_sample_sd_sd = 2,
  epsilon_mean = 0,
  epsilon_sd = 1,
  merge_matrix_files = NULL,
  merge_meta_files = NULL,
  keep_all_genes = FALSE,
  mat = NULL,
  raw.counts = FALSE,
  raw.norm.method = NULL,
  log2 = TRUE,
  quantnorm = FALSE,
  remove.duplicates = TRUE,
  meta = NULL,
  correction_method = "all",
  batch.1 = NULL,
  batch.2 = NULL,
  log2_transformed = TRUE,
  variable_of_interest = NULL,
  housekeeping = BatchFLEX::hsiao_mouse,
  k = 2,
  drop = 0,
  center = FALSE,
  round = FALSE,
  tolerance = 1e-08,
  par.prior = TRUE,
  sva_nsv_method = "be",
  evaluation_method = "all",
  annotation = "all",
  cluster_number = NULL,
  cluster_analysis_method = "all",
  color_by = "all",
  ncomponents = 5,
  pca_factors = NULL,
  variable_choices = NULL,
  plot_title = NULL,
  large_list = NULL,
  topN = 2000,
  var_type = "MAD",
  cluster_method = "ward.D2",
  fill_percentage = TRUE,
  heatmap_annotation = NULL,
  heatmap_rownames = FALSE,
  heatmap_colnames = FALSE,
  topN_pvca = 20000,
  pvca_pct = 0.8,
  cores = 1,
  parallelize = FALSE
)
}
\arguments{
\item{Batch_FLEX_function}{A Character vector of BatchFLEX functions in c("retrieve_data", "merge_data", "simulate_data", "preprocess_data", "batch_correct", "batch_evaluate").}

\item{num_samples}{An integer setting the number of samples to simulate.}

\item{num_genes}{An integer setting the number of genes to simulate.}

\item{num_control}{An integer setting the number of genes that do not respond to treatment.}

\item{num_treatments}{An integer setting the number of total treatments.}

\item{treatment_effect_multiplier}{A number or numbers setting how each treatment impacts the treatment effect. Must equal the number of treatments.}

\item{treatment_effect}{A number setting average impact of the treatments.}

\item{treatment_sd}{A number setting the standard deviation of the treatments.}

\item{additional_batch}{An integer setting the number of additional batches. e.g. setting to 1 indicates 2 total batches in the simulation.}

\item{batch_proportion}{A number setting the proportion of the total sample that is a different batch. Each additional batch is given an equal proportion of the total batch proportion.}

\item{batch_effect_multiplier}{A number or numbers setting how each batch impacts the treatment effect.}

\item{batch_effect}{A number setting the initial expression of each gene for a single sample for the batch effect. e.g. a single column of num_genes length that will be used to set the average expression for that gene across a set of batch samples}

\item{batch_effect_sd}{A number setting the deviation for the initial expression of each gene for a single sample for the batch effect.}

\item{batch_sample_sd_mean}{A number setting average for the standard deviation for each sample derived from the initial gene expression.}

\item{batch_sample_sd_sd}{A number setting the deviation for the average standard deviation used in the batch effect calculations.}

\item{epsilon_mean}{A number setting the average noise level}

\item{epsilon_sd}{A number setting the deviation for the average noise level}

\item{merge_matrix_files}{A list of matrix files to be merged. If supplied as a named list, batchflex will add a named column to the meta file with the specified names. If not, batch flex will add a generic name per meta file.}

\item{merge_meta_files}{A list of meta files to be merged.}

\item{keep_all_genes}{A TRUE/FALSE statement to indicate how unmatched genes will be handled. If TRUE, then NAs will be inserted for unmatched genes. If FALSE, unmatched genes will be deleted. Default is set to FALSE.}

\item{mat}{A Numeric matrix or list of matrices after pre-processing and/or batch correction with features as rownames and sample names as the column names.}

\item{raw.counts}{Logical. TRUE indicates the input data is raw counts. FALSE indicated the input data is not raw counts.}

\item{raw.norm.method}{Character string of what method to use for raw count normalization. Supported methods are "TMM" or "upperquartile". If raw.counts is FALSE this will be ignored.}

\item{log2}{Logical. If TRUE, the input data with be logged with the method "log2+1".}

\item{quantnorm}{Logical. If TRUE, the input data will be quantile normalized.}

\item{remove.duplicates}{Logical. If TRUE, if duplicate row features are found, they will be summarized to the row with the maximum average feature value.}

\item{meta}{Data frame of sample data with the first column being sample names that match the column names of the matrix.}

\item{correction_method}{A character vector of batch correction methods in c("Limma", "ComBat", "Mean Centering", "ComBatseq", "Harman", "RUVg", "SVA). Default is set to "all", which excludes ComBatseq as it requires a count matrix.}

\item{batch.1}{Column name from the meta file of the column that will be used for batch information.}

\item{batch.2}{Column name from the meta file of the column that will be used for batch two information.}

\item{log2_transformed}{logical whether the data is already transformed. Default is set to TRUE.}

\item{variable_of_interest}{Column name from the meta file of the column that will be used for the variable of interest information.}

\item{housekeeping}{Name of housekeeping gene set or character vector of housekeeping genes. Default is set to the preloaded HSIAO human gene list.}

\item{k}{Used in the RUVg correction_method, the number of factors of unwanted variation to be estimated from the data.}

\item{drop}{Used in the RUVg correction_method, the number of singular values to drop in the estimation of the factors of unwanted variation. This number is usually zero, but might be set to one if the first singular value captures the effect of interest. It must be less than k.}

\item{center}{Used in the RUVg correction_method, if TRUE, the counts are centered, for each gene, to have mean zero across samples. This is important to ensure that the first singular value does not capture the average gene expression.}

\item{round}{Used in the RUVg correction_method, if TRUE, the normalized measures are rounded to form pseudo-counts.}

\item{tolerance}{Used in the RUVg correction_method, tolerance in the selection of the number of positive singular values, i.e., a singular value must be larger than tolerance to be considered positive.}

\item{par.prior}{Used in the ComBat correction_method, TRUE indicates parametric adjustments will be used, FALSE indicates non-parametric adjustments will be used.}

\item{sva_nsv_method}{Input correction_method for the num.sv function in sva. Default is set to "be", but can be manually set to "leek".}

\item{evaluation_method}{A character vector of batch correction methods in c("pca", "cluster_analysis", "mc_pca", "pca_details", "rle", "ev", "sva", "umap").}

\item{annotation}{Used by evaluation pca to select whether a cluster or meta annotated PCA plot is generated. cluster = "cluster", meta = "meta", all = c("cluster", "meta").}

\item{cluster_number}{Used by evaluation pca and evaluation cluster HE to select the number of kmeans generated clusters to display. If NULL is selected, a Silhouette generated cluster number is used..}

\item{cluster_analysis_method}{Used to select cluster analysis method. Elbow = "wss", Silhouette = "silhouette", Dunn = "dunn', all generates plots from each method.}

\item{color_by}{Used by evaluation multiple components, pca, and rle to select which feature will be used to color the individuals. Choices are "batch", "variable_of_interest", "BnW", and "all". Default is set to "all".}

\item{ncomponents}{Used by evaluation multiple components to select the number of principal components that will be plotted. Default is set to 5.}

\item{pca_factors}{Used by evaluation pca details to select the Column name from the meta file of the column that will be used to group the summary details of the PCA plot. If NULL, batch is selected.}

\item{variable_choices}{Used by the explanatory variables function to select which variables to plot. Default is a combination of the batch and variable of interest.}

\item{plot_title}{A list of titles for each matrix submitted. Used to named the plots generated by the batch evaluate function. Default will generate titles based on the uncorrected and corrected matrices.}

\item{large_list}{A BatchFLEX generated list used with the export function.}

\item{topN}{A positive integer indicating the top number of genes to select based on variance. Default is set to 2000.}

\item{var_type}{The variance measure used to select the top genes based on variance. Default is set to MAD. Choices are "MAD", "CV", or "VAR".}

\item{cluster_method}{The cluster method used by hclust. Default is set to ward.D2. Choices are "ward.D2", "ward.D", "complete", "single", "average", "mccquitty", "median", or "centroid".}

\item{fill_percentage}{A TRUE/FALSE statement that determines if the y axis on the barplot will be a percentage or count. Default is set to TRUE}

\item{heatmap_annotation}{A vector of column names from the meta file to include in the heatmap annotation. Default is set to batch.1 and the variable of interest if nothing is provided.}

\item{heatmap_rownames}{A TRUE/FALSE statement to determine whether or not rownames will be included in the heatmap. Default is set to FALSE.}

\item{heatmap_colnames}{A TRUE/FALSE statement to determine whether or not colnames will be included in the heatmap. Default is set to FALSE.}

\item{topN_pvca}{A positive integer indicating the top number of genes to select based on variance for PVCA. Default is set to 20000.}

\item{pvca_pct}{A numeric between 0 and 1 that indicates the percentile value of the minimum amount of the variabilities that the selected principal components need to explain. Default is set to 0.8.}

\item{cores}{A numeric input for the number of cores to use if parallel is set to TRUE. Default is set to 1. If none is chosen and parallelize is TRUE default is set to the number of evaluation methods chosen, if possible.}

\item{parallelize}{A logical operator to determine if correction will be run sequentially or in parallel.}
}
\value{
A list object containing all plots and matrices generated by BatchFLEX.
}
\description{
Parent function of \code{BatchFLEX} used to preprocess data, simulate data, merge data, batch correct, batch evaluate, export data, and analyze data using a Shiny application.
}
\examples{
set.seed(333)
test_Batch_FLEX <- Batch_FLEX(mat = example_mat_abbreviated, meta = example_meta_abbreviated, batch.1 = "batchflex_study", variable_of_interest = "Major_Lineage")

}
