% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchFLEX_export.R
\name{BatchFLEX_export}
\alias{BatchFLEX_export}
\title{BatchFLEX_export}
\usage{
BatchFLEX_export(large_list)
}
\arguments{
\item{large_list}{A large list of plots and matrices generated by BatchFLEX}
}
\value{
A message depicting output location
}
\description{
BatchFLEX_export
}
\examples{
set.seed(333)
}
