% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchFLEX_shiny.R
\name{BatchFLEX_shiny}
\alias{BatchFLEX_shiny}
\title{BatchFLEX_shiny}
\usage{
BatchFLEX_shiny()
}
\value{
Message indicating BatchFLEX is starting
}
\description{
Used to run a shiny version of \code{Batch_FLEX} locally.
}
\examples{
set.seed(333)
BatchFLEX_shiny()

}
