% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eisenberg}
\alias{eisenberg}
\title{Eisenberg's housekeeping genes}
\format{
\subsection{\code{eisenberg}}{

a vector with house keeping genes
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/23810203/}
}
\usage{
eisenberg
}
\description{
Eisenberg's housekeeping genes
}
\keyword{datasets}
