% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{eisenberg_mouse}
\alias{eisenberg_mouse}
\title{Eisenberg's mouse housekeeping genes}
\format{
\subsection{\code{eisenberg_mouse}}{

a vector with house keeping genes
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/23810203/}
}
\usage{
eisenberg_mouse
}
\description{
Eisenberg's mouse housekeeping genes
}
\keyword{datasets}
