% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_cluster_analysis.R
\name{evaluation_cluster_analysis}
\alias{evaluation_cluster_analysis}
\title{Evaluation Cluster Analysis}
\usage{
evaluation_cluster_analysis(mat, cluster_analysis_method, plot_title)
}
\arguments{
\item{mat}{A Numeric matrix or list of matrices after pre-processing and/or batch correction with features as rownames and sample names as the column names}

\item{cluster_analysis_method}{Used to select cluster analysis method. Elbow = "wss", Silhouette = "silhouette", Dunn = "dunn', all generates plots from each method}

\item{plot_title}{A list of titles for each matrix submitted. Used to named the plots generated by the batch evaluate function. Default will generate titles based on the uncorrected and corrected matrices.}
}
\value{
A list object of cluster plots for each selected cluster analysis method
}
\description{
Evaluation Cluster Analysis
}
\examples{
set.seed(333)
}
