% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_cluster_HE.R
\name{evaluation_cluster_HE}
\alias{evaluation_cluster_HE}
\title{Evaluation Cluster Heterogeneity and Evenness}
\usage{
evaluation_cluster_HE(
  mat,
  meta,
  batch.1,
  cluster_number,
  topN,
  cluster_method,
  var_type,
  fill_percentage,
  plot_title
)
}
\arguments{
\item{mat}{A Numeric matrix or list of matrices after pre-processing and/or batch correction with features as rownames and sample names as the column names.}

\item{meta}{Data frame of sample data with the first column being sample names that match the column names of the matrix.}

\item{batch.1}{Column name from the meta file of the column that will be used for batch information.}

\item{cluster_number}{Used by evaluation pca to select the number of kmeans generated clusters to display in the plot. If NULL is selected, a Silhouette generated cluster number is used.}

\item{topN}{A positive integer indicating the top number of genes to select based on variance. Default is set to 2000.}

\item{cluster_method}{The cluster method used by hclust. Default is set to ward.D2. Choices are "ward.D2", "ward.D", "complete", "single", "average", "mccquitty", "median", or "centroid".}

\item{var_type}{The variance measure used to select the top genes based on variance. Default is set to MAD. Choices are "MAD", "CV", or "VAR".}

\item{fill_percentage}{A TRUE/FALSE statement that determines if the y axis on the barplot will be a percentage or count. Default is set to TRUE}

\item{plot_title}{A list of titles for each matrix submitted. Used to named the plots generated by the batch evaluate function. Default will generate titles based on the uncorrected and corrected matrices.}
}
\value{
A list object of cluster HE plots and matrices.
}
\description{
Evaluation Cluster Heterogeneity and Evenness
}
\examples{
set.seed(333)
}
