% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_ev.R
\name{evaluation_ev}
\alias{evaluation_ev}
\title{Evaluation Explanatory Variables}
\usage{
evaluation_ev(
  mat,
  meta,
  variable_choices,
  batch.1,
  variable_of_interest,
  plot_title
)
}
\arguments{
\item{mat}{A Numeric matrix or list of matrices after pre-processing and/or batch correction with features as rownames and sample names as the column names}

\item{meta}{Data frame of sample data with the first column being sample names that match the column names of the matrix}

\item{variable_choices}{Used by the explanatory variables function to select which variables to plot. Default is a combination of the batch and variable of interest}

\item{batch.1}{Column name from the meta file of the column that will be used for batch information}

\item{variable_of_interest}{Column name from the meta file of the column that will be used for the variable of interest information}
}
\value{
A list object of explanatory variables plots
}
\description{
Evaluation Explanatory Variables
}
\examples{
set.seed(333)
}
