% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_heatmap.R
\name{evaluation_heatmap}
\alias{evaluation_heatmap}
\title{Evaluation Heatmap}
\usage{
evaluation_heatmap(
  mat,
  meta,
  batch.1,
  variable_of_interest,
  topN,
  cluster_method,
  var_type,
  heatmap_annotation,
  heatmap_rownames,
  heatmap_colnames,
  plot_title
)
}
\arguments{
\item{mat}{A Numeric matrix or list of matrices after pre-processing and/or batch correction with features as rownames and sample names as the column names.}

\item{meta}{Data frame of sample data with the first column being sample names that match the column names of the matrix.}

\item{batch.1}{Column name from the meta file of the column that will be used for batch information.}

\item{variable_of_interest}{Column name from the meta file of the column that will be used for the variable of interest information.}

\item{topN}{A positive integer indicating the top number of genes to select based on variance. Default is set to 2000.}

\item{cluster_method}{he cluster method used by hclust. Default is set to ward.D2. Choices are "ward.D2", "ward.D", "complete", "single", "average", "mccquitty", "median", or "centroid".}

\item{var_type}{The variance measure used to select the top genes based on variance. Default is set to MAD. Choices are "MAD", "CV", or "VAR".}

\item{heatmap_annotation}{A vector of column names from the meta file to include in the heatmap annotation}

\item{heatmap_rownames}{A TRUE/FALSE statement to determine whether or not rownames will be included in the heatmap}

\item{heatmap_colnames}{A TRUE/FALSE statement to determine whether or not colnames will be included in the heatmap}

\item{plot_title}{A list of titles for each matrix submitted. Used to named the plots generated by the batch evaluate function. Default will generate titles based on the uncorrected and corrected matrices.}
}
\value{
A list object of heatmap plots.
}
\description{
Generates an annotated heatmap from a meta and expression matrix file using \code{ComplexHeatmap}.
}
\examples{
set.seed(333)
test_evaluate <- batch_evaluate(evaluation_method = "heatmap", mat = BatchFLEX::preprocess_matrix(BatchFLEX::example_mat), meta = example_meta, batch.1 = "batchflex_study", variable_of_interest = "Major_Lineage")
test_evaluate$Plots$heatmap

}
