% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_pca.R
\name{evaluation_pca}
\alias{evaluation_pca}
\title{Evaluation Pincipal Component Analysis}
\usage{
evaluation_pca(
  mat,
  meta,
  annotation,
  cluster_number,
  batch.1,
  variable_of_interest,
  color_by,
  plot_title
)
}
\arguments{
\item{mat}{A Numeric matrix or list of matrices after pre-processing and/or batch correction with features as rownames and sample names as the column names}

\item{meta}{Data frame of sample data with the first column being sample names that match the column names of the matrix}

\item{annotation}{Used to select whether a cluster or meta annotated PCA plot is generated. cluster = "cluster", meta = "meta", all = c("cluster", "meta")}

\item{cluster_number}{Used to select the number of kmeans generated clusters to display in the uncorrected plot. If NULL is selected, a Silhouette generated cluster number is used.}

\item{batch.1}{Column name from the meta file of the column that will be used for batch information}

\item{variable_of_interest}{Column name from the meta file of the column that will be used for the variable of interest information}

\item{color_by}{Used by evaluation multiple components, pca, and rle to select which feature will be used to color the individuals. Choices are "batch", "variable_of_interest", "BnW", and "all". Default is set to "all"}

\item{plot_title}{A list of titles for each matrix submitted. Used to named the plots generated by the batch evaluate function. Default will generate titles based on the uncorrected and corrected matrices.}
}
\value{
A list object of PCA plots for each annotation method selected
}
\description{
Generates an annotated PCA plot from a meta and expression matrix using \code{stats::prcomp} and \code{ggplot2}.
}
\examples{
set.seed(333)
test_evaluate <- batch_evaluate(evaluation_method = "pca", mat = BatchFLEX::preprocess_matrix(BatchFLEX::example_mat), meta = example_meta, batch.1 = "batchflex_study", variable_of_interest = "Major_Lineage", annotation = "meta")
test_evaluate$Plots$pca

}
