% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_pca_details.R
\name{evaluation_pca_details}
\alias{evaluation_pca_details}
\title{Evaluation PCA Details}
\usage{
evaluation_pca_details(mat, meta, batch.1, pca_factors, plot_title)
}
\arguments{
\item{mat}{A Numeric matrix or list of matrices after pre-processing and/or batch correction with features as rownames and sample names as the column names}

\item{meta}{Data frame of sample data with the first column being sample names that match the column names of the matrix}

\item{batch.1}{Column name from the meta file of the column that will be used for batch information}

\item{pca_factors}{Column name from the meta file of the column that will be used to group the summary details of the PCA plot. If NULL, batch is selected}

\item{plot_title}{A list of titles for each matrix submitted. Used to named the plots generated by the batch evaluate function. Default will generate titles based on the uncorrected and corrected matrices.}
}
\value{
A list object of PCA details including a scree plot, a matrix of pca components, a matrix of contributions grouped by factor, and a matrix of contribution counts
}
\description{
Evaluation PCA Details
}
\examples{
set.seed(333)
}
