% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_pvca.R
\name{evaluation_pvca}
\alias{evaluation_pvca}
\title{Evaluation PVCA}
\usage{
evaluation_pvca(
  mat,
  meta,
  batch.1,
  variable_of_interest,
  topN_pvca,
  var_type,
  variable_choices,
  pvca_pct,
  plot_title
)
}
\arguments{
\item{mat}{A Numeric matrix or list of matrices after pre-processing and/or batch correction with features as rownames and sample names as the column names.}

\item{meta}{Data frame of sample data with the first column being sample names that match the column names of the matrix.}

\item{batch.1}{Column name from the meta file of the column that will be used for batch information.}

\item{variable_of_interest}{Column name from the meta file of the column that will be used for the variable of interest information.}

\item{topN_pvca}{A positive integer indicating the top number of genes to select based on variance for PVCA. Default is set to 20000.}

\item{var_type}{The variance measure used to select the top genes based on variance. Default is set to MAD. Choices are "MAD", "CV", or "VAR".}

\item{variable_choices}{Used by the PVCA function to select which variables to plot. Default is a combination of the batch and variable of interest.}

\item{pvca_pct}{A numeric between 0 and 1 that indicates the percentile value of the minimum amount of the variabilities that the selected principal components need to explain.}

\item{plot_title}{A list of titles for each matrix submitted. Used to named the plots generated by the batch evaluate function. Default will generate titles based on the uncorrected and corrected matrices.}
}
\value{
A list object of PVCA plots.
}
\description{
Generates a bar plot based on PVCA calculations from a meta and expression matrix using \code{statVisual::PVCA}.
}
\examples{
set.seed(333)
test_evaluate <- batch_evaluate(evaluation_method = "pvca", mat = BatchFLEX::preprocess_matrix(BatchFLEX::example_mat), meta = example_meta, batch.1 = "batchflex_study", variable_of_interest = "Major_Lineage")
test_evaluate$Plots$pvca

}
