% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_sva.R
\name{evaluation_sva}
\alias{evaluation_sva}
\title{Evaluation Surrogate Variable Analysis}
\usage{
evaluation_sva(mat, meta, variable_of_interest, sva_nsv_method, plot_title)
}
\arguments{
\item{mat}{A Numeric matrix or list of matrices after pre-processing and/or batch correction with features as rownames and sample names as the column names}

\item{meta}{Data frame of sample data with the first column being sample names that match the column names of the matrix}

\item{variable_of_interest}{Column name from the meta file of the column that will be used for the variable of interest information}

\item{sva_nsv_method}{Used by SVA to generate surrogate variables. Choices are "be" and "leek". Default is set to "be"}

\item{plot_title}{A list of titles for each matrix submitted. Used to named the plots generated by the batch evaluate function. Default will generate titles based on the uncorrected and corrected matrices.}
}
\value{
A list object of probability association plots and surrogate variable matrices
}
\description{
Generates a density plot (\code{ggplot2}) of associated probabilities calculated from \code{sva} using a meta and expression matrix.
Also generates a matrix of surrogate variables calculated from \code{sva}.
}
\examples{
set.seed(333)
test_evaluate <- batch_evaluate(evaluation_method = "sva", mat = BatchFLEX::preprocess_matrix(BatchFLEX::example_mat), meta = example_meta, batch.1 = "batchflex_study", variable_of_interest = "Major_Lineage", sva_nsv_method = "be")
test_evaluate$Plots$sva

}
