% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_umap.R
\name{evaluation_umap}
\alias{evaluation_umap}
\title{Evaluation umap}
\usage{
evaluation_umap(mat, meta, batch.1, variable_of_interest, color_by)
}
\arguments{
\item{mat}{A Numeric matrix or list of matrices after pre-processing and/or batch correction with features as rownames and sample names as the column names}

\item{meta}{Data frame of sample data with the first column being sample names that match the column names of the matrix}

\item{batch.1}{Column name from the meta file of the column that will be used for batch information}

\item{variable_of_interest}{Column name from the meta file of the column that will be used for the variable of interest information}

\item{color_by}{Used by evaluation multiple components, pca, and rle to select which feature will be used to color the individuals. Choices are "batch", "variable_of_interest", and "all". Default is set to "all"}
}
\value{
A list object of UMAP plots for each annotation method selected
}
\description{
Evaluation umap
}
\examples{
set.seed(333)
}
