% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_mat}
\alias{example_mat}
\title{Example expression data}
\format{
\subsection{\code{example_mat}}{

a numeric matrix containing transformed values with samples in columns and gene names in the rows
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE112876}

\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE15907}

\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE37448}

\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE60336}

\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE60337}

\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75202}
}
\usage{
example_mat
}
\description{
Example expression data
}
\keyword{datasets}
