% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{example_meta}
\alias{example_meta}
\title{Example meta data}
\format{
\subsection{\code{example_meta}}{

a data frame containing sample level information line name and study
\describe{
\item{Name}{sample name}
\itme{Study}{GSE study that the sampel came from}
\item{CellType}{cell type that was assigned to the sample}
}
}
}
\usage{
example_meta
}
\description{
Example meta data
}
\keyword{datasets}
