% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE112876_matrix}
\alias{GSE112876_matrix}
\title{GSE112876_matrix}
\format{
\subsection{\code{GSE112876_matrix}}{
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE112876}
}
\usage{
GSE112876_matrix
}
\description{
GSE112876_matrix
}
\keyword{datasets}
