% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE15907_matrix}
\alias{GSE15907_matrix}
\title{GSE15907_matrix}
\format{
\subsection{\code{GSE15907_matrix}}{
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE15907}
}
\usage{
GSE15907_matrix
}
\description{
GSE15907_matrix
}
\keyword{datasets}
