% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE37448_matrix}
\alias{GSE37448_matrix}
\title{GSE37448_matrix}
\format{
\subsection{\code{GSE37448_matrix}}{
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE37448}
}
\usage{
GSE37448_matrix
}
\description{
GSE37448_matrix
}
\keyword{datasets}
