% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE37448_meta}
\alias{GSE37448_meta}
\title{GSE37448_meta}
\format{
\subsection{\code{GSE37448_meta}}{

a data frame containing sample level information
\describe{
\item{SampleID}{sample ID}
\item{Study}{GSE study that the sample came from}
\item{MajorCellType}{Major cell type that was assigned to the sample}
\item{Major Lineage}{Major lineage that was assigned to the sample}
\item{Cell Family}{Cell Family that was assigned to the sample}
\item{GEOname}{Unique GEO name for each sample}
\item{Organism}{Species sample was taken from}
\item{GeneticBackground}{Organism strain}
\item{Sex}{Sex of organism that the sample was taken from}
\item{Tissue Source}{Tissue source for the sample}
\item{Exclusion Marker}{Any exclusion markers used to differentiate cells}
\item{Treatment}{Treatment administered to the organism}
\item{CellType}{Specific cell type that was assigned to the sample}
\item{Phenotype_Marker}{Phenotypic marker used to identify the cell family}
\item{Separation Method}{Method of separation used for each sample}
}
}
}
\usage{
GSE37448_meta
}
\description{
GSE37448_meta
}
\keyword{datasets}
