% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE60336_matrix}
\alias{GSE60336_matrix}
\title{GSE60336_matrix}
\format{
\subsection{\code{GSE60336_matrix}}{
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE60336}
}
\usage{
GSE60336_matrix
}
\description{
GSE60336_matrix
}
\keyword{datasets}
