% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE75202_matrix}
\alias{GSE75202_matrix}
\title{GSE75202_matrix}
\format{
\subsection{\code{GSE75202_matrix}}{
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75202}
}
\usage{
GSE75202_matrix
}
\description{
GSE75202_matrix
}
\keyword{datasets}
