% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{GSE75203_matrix}
\alias{GSE75203_matrix}
\title{GSE75203_matrix}
\format{
\subsection{\code{GSE75203_matrix}}{
}

\subsection{\code{GSE75203_matrix}}{
}
}
\source{
\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75203}

\url{https://www.ncbi.nlm.nih.gov/geo/query/acc.cgi?acc=GSE75203}
}
\usage{
GSE75203_matrix

GSE75203_matrix
}
\description{
GSE75203_matrix

GSE75203_matrix
}
\keyword{datasets}
