% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hsiao_mouse}
\alias{hsiao_mouse}
\title{HSAIO's mouse housekeeping genes}
\format{
\subsection{\code{hsiao_mouse}}{

a vector with house keeping genes
}
}
\source{
\url{https://www.gsea-msigdb.org/gsea/msigdb/cards/HSIAO_HOUSEKEEPING_GENES}
}
\usage{
hsiao_mouse
}
\description{
HSAIO's mouse housekeeping genes
}
\keyword{datasets}
