% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lin500}
\alias{lin500}
\title{LIN500's housekeeping genes}
\format{
\subsection{\code{lin500}}{

a vector with house keeping genes
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/23810203/}
}
\usage{
lin500
}
\description{
LIN500's housekeeping genes
}
\keyword{datasets}
