% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{lin500_mouse}
\alias{lin500_mouse}
\title{LIN500's mouse housekeeping genes}
\format{
\subsection{\code{lin500_mouse}}{

a vector with house keeping genes
}
}
\source{
\url{https://pubmed.ncbi.nlm.nih.gov/23810203/}
}
\usage{
lin500_mouse
}
\description{
LIN500's mouse housekeeping genes
}
\keyword{datasets}
