% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_umaps.R
\name{plot_umaps}
\alias{plot_umaps}
\title{Plot UMAPs}
\usage{
plot_umaps(adjusted_list, meta, color = NULL, method = NULL)
}
\arguments{
\item{adjusted_list}{list of matrices from \code{batch_correction()}}

\item{meta}{data frame of the sample level information}

\item{color}{column in \code{meta} which to color points - usually batch}

\item{method}{character string for which batch corrections to plot}
}
\value{
list of plots including unadjusted data and adjusted data
}
\description{
Plot UMAPs
}
\examples{
set.seed(333)
}
