% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess_matrix.R
\name{preprocess_matrix}
\alias{preprocess_matrix}
\title{Pre Process Matrix}
\usage{
preprocess_matrix(
  data = NULL,
  raw.counts = FALSE,
  raw.norm.method = NULL,
  log2 = TRUE,
  quantnorm = TRUE,
  remove.duplicates = TRUE
)
}
\arguments{
\item{data}{Numeric matrix or data frame. The features can be rownames or the first column followed by the sample names as the column names with numeric count data as the values.}

\item{raw.counts}{Logical. TRUE indicates the input data is raw counts. FALSE indicated the input data is not raw counts.}

\item{raw.norm.method}{Character string of what method to use for raw count normalization. Supported methods are "TMM" or "upperquartile". If raw.counts is FALSE this will be ignored.}

\item{log2}{Logical. If TRUE, the input data with be logged with the method "log2+1".}

\item{quantnorm}{Logical. If TRUE, the input data will be quantile normalized.}

\item{remove.duplicates}{Logical. If TRUE, if duplicate row features are found, they will be summarized to the row with the maximum average feature value.}
}
\value{
A numeric matrix that has been processed through the users input choices.
}
\description{
Pre Process Matrix
}
\examples{
set.seed(101)
}
