% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_merge_files.R
\name{process_merge_files}
\alias{process_merge_files}
\title{Process Merge Files}
\usage{
process_merge_files(
  files_tobe_merged_path,
  meta_files_included = TRUE,
  all_transnormed = TRUE
)
}
\arguments{
\item{files_tobe_merged_path}{A path to the folder containing the matrix and meta files.}

\item{meta_files_included}{A TRUE/FALSE to determine if the function should attempt to group a matrix and meta file together.}

\item{all_transnormed}{<- A TRUE/FALSE to determine if all matrices have been transformed and normalized equally.}
}
\value{
A list of matrix and meta files in appropriate format to be merged using \code{merge_data}
}
\description{
Process Merge Files
}
\examples{
set.seed(333)

}
