% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_data.R
\name{simulate_data}
\alias{simulate_data}
\title{Simulate Data}
\usage{
simulate_data(
  num_samples = 100,
  num_genes = 10000,
  num_control = 9000,
  num_treatments = 2,
  treatment_effect_multiplier = c(0, 1),
  treatment_effect = 5,
  treatment_sd = 1,
  additional_batch = 1,
  batch_proportion = 0.5,
  batch_effect_multiplier = c(1),
  batch_effect = 5,
  batch_effect_sd = 1,
  batch_sample_sd_mean = 0,
  batch_sample_sd_sd = 2,
  epsilon_mean = 0,
  epsilon_sd = 1
)
}
\arguments{
\item{num_samples}{An integer setting the number of samples to simulate.}

\item{num_genes}{An integer setting the number of genes to simulate.}

\item{num_control}{An integer setting the number of genes that do not respond to treatment.}

\item{num_treatments}{An integer setting the number of total treatments.}

\item{treatment_effect_multiplier}{A number or numbers setting how each treatment impacts the treatment effect. Must equal the number of treatments.}

\item{treatment_effect}{A number setting average impact of the treatments.}

\item{treatment_sd}{A number setting the standard deviation of the treatments.}

\item{additional_batch}{An integer setting the number of additional batches. e.g. setting to 1 indicates 2 total batches in the simulation.}

\item{batch_proportion}{A number setting the proportion of the total sample that is a different batch. Each additional batch is given an equal proportion of the total batch proportion.}

\item{batch_effect_multiplier}{A number or numbers setting how each batch impacts the treatment effect.}

\item{batch_effect}{A number setting the initial expression of each gene for a single sample for the batch effect. e.g. a single column of num_genes length that will be used to set the average expression for that gene across a set of batch samples}

\item{batch_effect_sd}{A number setting the deviation for the initial expression of each gene for a single sample for the batch effect.}

\item{batch_sample_sd_mean}{A number setting average for the standard deviation for each sample derived from the initial gene expression.}

\item{batch_sample_sd_sd}{A number setting the deviation for the average standard deviation used in the batch effect calculations.}

\item{epsilon_mean}{A number setting the average noise level}

\item{epsilon_sd}{A number setting the deviation for the average noise level}

\item{BatchFLEX_function}{A Character vector of BatchFLEX functions in c("retrieve_data", "simulate_data", "preprocess_data", "batch_correct", "batch_evaluate").}
}
\value{
A list of a simulated matrix and a simulated meta file
}
\description{
Generates simulated data based on the linear model framework introduced by Gagnon-Bartsch and Speed.
}
\examples{
set.seed(333)
simulated_data = simulate_data()
head(as.data.frame(simulated_data$sim_matrix), n = c(5,5))
head(simulated_data$sim_meta, n = c(5,5))
test_simulate_BatchFLEX <- Batch_FLEX(Batch_FLEX_function = c("batch_correct", "batch_evaluate"), correction_method = "ComBat", evaluation_method = "pca", mat = simulated_data$sim_matrix, meta = simulated_data$sim_meta, batch.1 = "batch", color_by = "batch")
}
