% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_combatseq.R
\name{adjust_combatseq}
\alias{adjust_combatseq}
\title{Adjust ComBat_Seq}
\usage{
adjust_combatseq(mat, meta, batch.1, treatment, log2_transformed)
}
\arguments{
\item{mat}{Numeric matrix with features as rownames and sample names as the column names}

\item{meta}{Data frame of sample data with the first column being sample names that match the column names of the matrix}

\item{batch.1}{Column name from the meta file of the column that will be used for batch one information}

\item{treatment}{Column name from the meta file of the column that will be used for treatment information}

\item{log2_transformed}{logical whether the data is alrady transformed}
}
\value{
an adjusted expression matrix
}
\description{
Wrapper for \code{sva::ComBat_seq}
}
\examples{
set.seed(333)
}
