% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_combatseq.R
\name{adjust_combatseq}
\alias{adjust_combatseq}
\title{Adjust ComBat_Seq}
\usage{
adjust_combatseq(mat, meta, batch.1, variable_of_interest, log2_transformed)
}
\arguments{
\item{mat}{Numeric matrix with features as rownames and sample names as the column names}

\item{meta}{Data frame of sample data with the first column being sample names that match the column names of the matrix}

\item{batch.1}{Column name from the meta file of the column that will be used for batch one information}

\item{variable_of_interest}{Column name from the meta file of the column that will be used for variable_of_interest information}

\item{log2_transformed}{logical whether the data is alrady transformed}
}
\value{
an adjusted expression matrix
}
\description{
Remove batch effects from expression data using CombatSeq from \code{sva}.
}
\examples{
set.seed(333)
adjusted_data <-  batch_correct(mat = BatchFLEX::preprocess_matrix(BatchFLEX::example_mat),
meta = BatchFLEX::example_meta,
correction_method = "ComBatseq",
batch.1 = "batchflex_study",
variable_of_interest = "Major_Lineage")
head(as.data.frame(adjusted_data), n = c(5,5))

}
