% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_harman.R
\name{adjust_harman}
\alias{adjust_harman}
\title{Adjust Harmon}
\usage{
adjust_harman(mat, meta, batch.1, variable_of_interest, log2_transformed)
}
\arguments{
\item{mat}{Numeric matrix with features as rownames and sample names as the column names}

\item{meta}{Data frame of sample data with the first column being sample names that match the column names of the matrix}

\item{batch.1}{Column name from the meta file of the column that will be used for batch one information}

\item{variable_of_interest}{Column name from the meta file of the column that will be used for variable_of_interest information}

\item{log2_transformed}{logical whether the data is alrady transformed}
}
\value{
an adjusted expression matrix
}
\description{
Wrapper for \code{Harman::harman} and \code{Harman::reconstructData}
}
\examples{
set.seed(333)
}
