% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adjust_mean_centering.R
\name{adjust_mean_centering}
\alias{adjust_mean_centering}
\title{Adjust Mean Centering}
\usage{
adjust_mean_centering(mat, meta, batch.1, log2_transformed)
}
\arguments{
\item{mat}{Numeric matrix with features as rownames and sample names as the column names}

\item{meta}{Data frame of sample data with the first column being sample names that match the column names of the matrix}

\item{batch.1}{Column name from the meta file of the column that will be used for batch one information}

\item{log2_transformed}{whether data is log2 transformed already or not}
}
\value{
an adjusted expression matrix
}
\description{
Wrapper for \code{bapred::meancenter}
}
\examples{
set.seed(333)
}
