% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_correct.R
\name{batch_correct}
\alias{batch_correct}
\title{Batch Correct}
\usage{
batch_correct(
  mat = NULL,
  meta = NULL,
  correction_method = "all",
  batch.1 = NULL,
  batch.2 = NULL,
  log2_transformed = TRUE,
  variable_of_interest = NULL,
  housekeeping = BatchFLEX::hsiao_mouse,
  k = 2,
  drop = 0,
  center = FALSE,
  round = FALSE,
  tolerance = 1e-08,
  par.prior = TRUE,
  sva_nsv_method = "be"
)
}
\arguments{
\item{mat}{A Numeric matrix after preprocessing with features as rownames and sample names as the column names.}

\item{meta}{Data frame of sample data with the first column being sample names that match the column names of the matrix.}

\item{correction_method}{A character vector of batch correction methods in c("Limma", "ComBat", "Mean Centering", "ComBatseq", "Harman", "RUVg", "SVA).}

\item{batch.1}{Column name from the meta file of the column that will be used for batch one information.}

\item{batch.2}{Column name from the meta file of the column that will be used for batch two information.}

\item{log2_transformed}{logical whether the data is already transformed. Default is set to TRUE.}

\item{variable_of_interest}{Column name from the meta file of the column that will be used for variable of interest information.}

\item{housekeeping}{Name of housekeeping gene set or character vector of housekeeping genes.}

\item{k}{Used in the RUVg correction_method, the number of factors of unwanted variation to be estimated from the data.}

\item{drop}{Used in the RUVg correction_method, the number of singular values to drop in the estimation of the factors of unwanted variation. This number is usually zero, but might be set to one if the first singular value captures the effect of interest. It must be less than k.}

\item{center}{Used in the RUVg correction_method, if TRUE, the counts are centered, for each gene, to have mean zero across samples. This is important to ensure that the first singular value does not capture the average gene expression.}

\item{round}{Used in the RUVg correction_method, if TRUE, the normalized measures are rounded to form pseudo-counts.}

\item{tolerance}{Used in the RUVg correction_method, tolerance in the selection of the number of positive singular values, i.e., a singular value must be larger than tolerance to be considered positive.}

\item{par.prior}{Used in the ComBat correction_method, TRUE indicates parametric adjustments will be used, FALSE indicates non-parametric adjustments will be used.}

\item{sva_nsv_method}{Input correction_method for the num.sv function in sva. Default is set to "be", but can be manually set to "leek".}

\item{prep_matrix}{Logical. If TRUE, run preprocess_matrix function or log and/or quantile normalize or normalize raw counts.}

\item{raw.counts}{Logical. TRUE indicates the input data is raw counts. FALSE indicated the input data is not raw counts.}

\item{raw.norm.method}{Character string of what method to use for raw count normalization. Supported methods are "TMM" or "upperquartile". If raw.counts is FALSE this will be ignored.}

\item{log2}{Logical. If TRUE, the input data with be logged with the method "log2+1".}

\item{quantnorm}{Logical. If TRUE, the input data will be quantile normalized.}

\item{remove.duplicates}{Logical. If TRUE, if duplicate row features are found, they will be summarized to the row with the maximum average feature value.}
}
\value{
List object of length of correction_method
}
\description{
Batch Correct
}
\examples{
set.seed(333)
}
