% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/batch_evaluate.R
\name{batch_evaluate}
\alias{batch_evaluate}
\title{Batch Evaluate}
\usage{
batch_evaluate(
  mat = NULL,
  meta = NULL,
  evaluation_method = "all",
  batch.1 = NULL,
  annotation = "all",
  cluster_number = NULL,
  variable_of_interest = NULL,
  cluster_analysis_method = "all",
  color_by = "all",
  ncomponents = 5,
  pca_factors = NULL,
  variable_choices = NULL,
  sva_nsv_method = "be",
  plot_title = NULL,
  topN = 2000,
  var_type = "MAD",
  cluster_method = "ward.D2",
  fill_percentage = TRUE,
  heatmap_annotation = NULL,
  heatmap_rownames = FALSE,
  heatmap_colnames = FALSE,
  pvca_pct = 0.8
)
}
\arguments{
\item{mat}{A Numeric matrix or list of matrices after pre-processing and/or batch correction with features as rownames and sample names as the column names.}

\item{meta}{Data frame of sample data with the first column being sample names that match the column names of the matrix.}

\item{evaluation_method}{A character vector of batch correction methods in c("pca", "cluster_analysis", "mc_pca", "pca_details", "rle", "ev", "sva", "umap").}

\item{batch.1}{Column name from the meta file of the column that will be used for batch information.}

\item{annotation}{Used by evaluation pca to select whether a cluster or meta annotated PCA plot is generated. cluster = "cluster", meta = "meta", all = c("cluster", "meta").}

\item{cluster_number}{Used by evaluation pca and evaluation cluster HE to select the number of kmeans generated clusters to display in the plot. If NULL is selected, a Silhouette generated cluster number is used.}

\item{variable_of_interest}{Column name from the meta file of the column that will be used for the variable of interest information}

\item{cluster_analysis_method}{Used by evaluation cluster analysis to select cluster analysis method. Elbow = "wss", Silhouette = "silhouette", Dunn = "dunn', all generates plots from each method.}

\item{color_by}{Used by evaluation multiple components, pca, and rle to select which feature will be used to color the individuals. Choices are "batch", "variable_of_interest", and "all". Default is set to "all".}

\item{ncomponents}{Used by evaluation multiple components to select the number of principal components that will be plotted. Default is set to 5.}

\item{pca_factors}{Used by evaluation pca details to select the Column name from the meta file of the column that will be used to group the summary details of the PCA plot. If NULL, batch is selected.}

\item{variable_choices}{Used by evaluation explanatory variables to select which variables to plot. Default is a combination of the batch and variable of interest.}

\item{sva_nsv_method}{Used by evaluation SVA to generate surrogate variables. Choices are "be" and "leek". Default is set to "be".}

\item{plot_title}{A list of titles for each matrix submitted. Used to named the plots generated by the batch evaluate function. Default will generate titles based on the uncorrected and corrected matrices.}

\item{topN}{A positive integer indicating the top number of genes to select based on variance. Default is set to 2000.}

\item{var_type}{The variance measure used to select the top genes based on variance. Default is set to MAD. Choices are "MAD", "CV", or "VAR".}

\item{cluster_method}{The cluster method used by hclust. Default is set to ward.D2. Choices are "ward.D2", "ward.D", "complete", "single", "average", "mccquitty", "median", or "centroid".}

\item{fill_percentage}{A TRUE/FALSE statement that determines if the y axis on the barplot will be a percentage or count. Default is set to TRUE}

\item{pvca_pct}{A numeric between 0 and 1 that indicates the percentile value of the minimum amount of the variabilities that the selected principal components need to explain. Default is set to 0.8.}
}
\value{
A list object separated by Plots and Matrices containing all plots and matrices generated by the selected evaluation methods
}
\description{
Batch Evaluate
}
\examples{
set.seed(333)
}
