% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Batch_FLEX.R
\name{Batch_FLEX}
\alias{Batch_FLEX}
\title{BatchFLEX}
\usage{
Batch_FLEX(
  Batch_FLEX_function = NULL,
  mat = NULL,
  meta = NULL,
  correction_method = NULL,
  batch.1 = NULL,
  batch.2 = NULL,
  log2_transformed = TRUE,
  variable_of_interest = NULL,
  housekeeping = NULL,
  k = 2,
  drop = 0,
  center = FALSE,
  round = FALSE,
  tolerance = 1e-08,
  par.prior = TRUE,
  sva_nsv_method = "be",
  batch_correction = NULL,
  evaluation_method = NULL,
  annotation = NULL,
  uncorrected_cluster_number = NULL,
  corrected_cluster_number = NULL,
  cluster_analysis_method = NULL,
  color_by = "batch",
  ncomponents = 5,
  pca_factors = NULL,
  variable_choices = NULL
)
}
\arguments{
\item{Batch_FLEX_function}{A Character vector of BatchFLEX functions in c("retrieve_data", "generate_data", "preprocess_data", "batch_correct", "batch_evaluate")}

\item{mat}{Numeric matrix after pre-processing with features as rownames and sample names as the column names}

\item{meta}{Data frame of sample data with the first column being sample names that match the column names of the matrix}

\item{correction_method}{A character vector of batch correction methods in c("Limma", "ComBat", "Mean Centering", "ComBatseq", "Harman", "RUVg", "SVA)}

\item{batch.1}{Column name from the meta file of the column that will be used for batch information}

\item{batch.2}{Column name from the meta file of the column that will be used for batch two information}

\item{log2_transformed}{logical whether the data is already transformed. Default is set to TRUE}

\item{variable_of_interest}{Column name from the meta file of the column that will be used for the variable of interest information}

\item{housekeeping}{Name of housekeeping gene set or character vector of housekeeping genes}

\item{k}{Used in the RUVg correction_method, the number of factors of unwanted variation to be estimated from the data}

\item{drop}{Used in the RUVg correction_method, the number of singular values to drop in the estimation of the factors of unwanted variation. This number is usually zero, but might be set to one if the first singular value captures the effect of interest. It must be less than k}

\item{center}{Used in the RUVg correction_method, if TRUE, the counts are centered, for each gene, to have mean zero across samples. This is important to ensure that the first singular value does not capture the average gene expression}

\item{round}{Used in the RUVg correction_method, if TRUE, the normalized measures are rounded to form pseudo-counts}

\item{tolerance}{Used in the RUVg correction_method, tolerance in the selection of the number of positive singular values, i.e., a singular value must be larger than tolerance to be considered positive}

\item{par.prior}{Used in the ComBat correction_method, TRUE indicates parametric adjustments will be used, FALSE indicates non-parametric adjustments will be used}

\item{sva_nsv_method}{Input correction_method for the num.sv function in sva. Default is set to "be", but can be manually set to "leek"}

\item{batch_correction}{Numeric matrix following batch correction with features as rownames and sample names as the column names. If NULL BatchFLEX will generate using batch_correct}

\item{evaluation_method}{A character vector of batch correction methods in c("pca", "cluster_analysis", "mc_pca", "pca_details", "rle", "ev", "sva")}

\item{annotation}{Used by evaluation pca to select whether a cluster or meta annotated PCA plot is generated. cluster = "cluster", meta = "meta", all = c("cluster", "meta")}

\item{uncorrected_cluster_number}{Used by evaluation pca to select the number of kmeans generated clusters to display in the uncorrected plot. If NULL is selected, a Dunn generated cluster number is used.}

\item{corrected_cluster_number}{Used by evaluation pca to select the number of kmeans generated clusters to display in the corrected plot. If NULL is selected, a Dunn generated cluster number is used.}

\item{cluster_analysis_method}{Used to select cluster analysis method. Elbow = "wss", Silhouette = "silhouette", Dunn = "dunn', all generates plots from each method}

\item{color_by}{Used by evaluation multiple components, pca, and rle to select which feature will be used to color the individuals}

\item{ncomponents}{Used by evaluation multiple components to select the number of principal components that will be plotted. Default is set to 5}

\item{pca_factors}{Used by evaluation pca details to select the Column name from the meta file of the column that will be used to group the summary details of the PCA plot. If NULL, batch is selected}

\item{variable_choices}{Used by the explanatory variables function to select which variables to plot. Default is a combination of the batch and variable of interest}
}
\value{
A list object containing all plots and matrices generated by BatchFLEX
}
\description{
BatchFLEX
}
\examples{
set.seed(333)
}
