% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchFLEX_export.R
\name{BatchFLEX_export}
\alias{BatchFLEX_export}
\title{BatchFLEX_export}
\usage{
BatchFLEX_export(large_list)
}
\arguments{
\item{large_list}{A large list of plots and matrices generated by BatchFLEX}
}
\value{
A message depicting output location
}
\description{
Export function for a \code{Batch_FLEX} generated list.
}
\examples{
set.seed(333)
test_Batch_FLEX <- Batch_FLEX(Batch_FLEX_function = c("preprocess_matrix", "batch_evaluate", "batch_correct", "BatchFLEX_export", "BatchFLEX_shiny"), correction_method = c("ComBat", "Limma", "Harman", "SVA", "RUVg"), mat = example_mat, meta = example_meta, batch.1 = "batchflex_study", variable_of_interest = "Major_Lineage")

}
