% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BatchFLEX_shiny.R
\name{BatchFLEX_shiny}
\alias{BatchFLEX_shiny}
\title{BatchFLEX_shiny}
\usage{
BatchFLEX_shiny()
}
\value{
Message indicating BatchFLEX is starting
}
\description{
BatchFLEX_shiny
}
\examples{
set.seed(333)
}
