% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_cluster_analysis.R
\name{evaluation_cluster_analysis}
\alias{evaluation_cluster_analysis}
\title{Evaluation Cluster Analysis}
\usage{
evaluation_cluster_analysis(mat, cluster_analysis_method, plot_title)
}
\arguments{
\item{mat}{A Numeric matrix or list of matrices after pre-processing and/or batch correction with features as rownames and sample names as the column names}

\item{cluster_analysis_method}{Used to select cluster analysis method. Elbow = "wss", Silhouette = "silhouette", Dunn = "dunn', all generates plots from each method}

\item{plot_title}{A list of titles for each matrix submitted. Used to named the plots generated by the batch evaluate function. Default will generate titles based on the uncorrected and corrected matrices.}
}
\value{
A list object of cluster plots for each selected cluster analysis method
}
\description{
Generates cluster analysis plots using wss, silhouette, and Dunn indexes calculated from expression matrices.
}
\examples{
set.seed(333)
test_evaluate <- batch_evaluate(evaluation_method = "cluster_analysis", mat = BatchFLEX::preprocess_matrix(BatchFLEX::example_mat), meta = example_meta, batch.1 = "batchflex_study", variable_of_interest = "Major_Lineage", cluster_analysis_method = "silhouette")
test_evaluate$Plots$cluster_analysis

}
