% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_mc_pca.R
\name{evaluation_mc_pca}
\alias{evaluation_mc_pca}
\title{Evaluation Multiple Components Plot}
\usage{
evaluation_mc_pca(
  mat,
  meta,
  batch.1,
  variable_of_interest,
  ncomponents,
  color_by,
  plot_title
)
}
\arguments{
\item{mat}{A Numeric matrix or list of matrices after pre-processing and/or batch correction with features as rownames and sample names as the column names}

\item{meta}{Data frame of sample data with the first column being sample names that match the column names of the matrix}

\item{batch.1}{Column name from the meta file of the column that will be used for batch information}

\item{variable_of_interest}{Column name from the meta file of the column that will be used for the variable of interest information}

\item{ncomponents}{Used by the multiple components plot to select the number of principal components that will be plotted. Default is set to 5}

\item{color_by}{Used by evaluation multiple components, pca, and rle to select which feature will be used to color the individuals. Choices are "batch", "variable_of_interest", "BnW", and "all". Default is set to "all"}

\item{plot_title}{A list of titles for each matrix submitted. Used to named the plots generated by the batch evaluate function. Default will generate titles based on the uncorrected and corrected matrices.}
}
\value{
A list object of multiple components plots
}
\description{
Generates a multiple components PCA plot from a meta and matrix file using \code{scater::plotPCA}.
}
\examples{
set.seed(333)
test_evaluate <- batch_evaluate(evaluation_method = "mc_pca", mat = BatchFLEX::preprocess_matrix(BatchFLEX::example_mat), meta = example_meta, batch.1 = "batchflex_study", variable_of_interest = "Major_Lineage", color_by = "batch", ncomponents = 10)
test_evaluate$Plots$mc_pca

}
