% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_mc_pca.R
\name{evaluation_mc_pca}
\alias{evaluation_mc_pca}
\title{Evaluation Multiple Components Plot}
\usage{
evaluation_mc_pca(
  mat,
  meta,
  batch.1,
  variable_of_interest,
  ncomponents,
  color_by
)
}
\arguments{
\item{mat}{Numeric matrix after pre-processing with features as rownames and sample names as the column names}

\item{meta}{Data frame of sample data with the first column being sample names that match the column names of the matrix}

\item{batch.1}{Column name from the meta file of the column that will be used for batch information}

\item{variable_of_interest}{Column name from the meta file of the column that will be used for the variable of interest information}

\item{ncomponents}{Used by the multiple components plot to select the number of principal components that will be plotted. Default is set to 5}

\item{color_by}{Used by evaluation multiple components, pca, and rle to select which feature will be used to color the individuals. Choices are "batch", "variable_of_interest", "BnW", and "all". Default is set to "all"}
}
\value{
A list object of multiple components plots
}
\description{
Evaluation Multiple Components Plot
}
\examples{
set.seed(333)
}
