% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_pca.R
\name{evaluation_pca}
\alias{evaluation_pca}
\title{Evaluation Pincipal Component Analysis}
\usage{
evaluation_pca(
  mat,
  batch_correction,
  meta,
  annotation,
  uncorrected_cluster_number,
  corrected_cluster_number,
  batch.1,
  variable_of_interest,
  color_by
)
}
\arguments{
\item{mat}{Numeric matrix after pre-processing with features as rownames and sample names as the column names}

\item{batch_correction}{Numeric matrix following batch correction with features as rownames and sample names as the column names}

\item{meta}{Data frame of sample data with the first column being sample names that match the column names of the matrix}

\item{annotation}{Used to select whether a cluster or meta annotated PCA plot is generated. cluster = "cluster", meta = "meta", all = c("cluster", "meta")}

\item{uncorrected_cluster_number}{Used to select the number of kmeans generated clusters to display in the uncorrected plot. If NULL is selected, a Dunn generated cluster number is used.}

\item{corrected_cluster_number}{Used to select the number of kmeans generated clusters to display in the corrected plot. If NULL is selected, a Dunn generated cluster number is used.}

\item{batch.1}{Column name from the meta file of the column that will be used for batch information}

\item{variable_of_interest}{Column name from the meta file of the column that will be used for the variable of interest information}

\item{color_by}{Used by the PCA plot to select which feature will be used to color the individuals}
}
\value{
A list object of uncorrected and batch corrected PCA plots for each annotation method selected
}
\description{
Evaluation Pincipal Component Analysis
}
\examples{
set.seed(333)
}
