% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_rle.R
\name{evaluation_rle}
\alias{evaluation_rle}
\title{Evaluation Relative Log Expression}
\usage{
evaluation_rle(
  mat,
  batch_correction,
  meta,
  batch.1,
  variable_of_interest,
  color_by
)
}
\arguments{
\item{mat}{Numeric matrix after pre-processing with features as rownames and sample names as the column names}

\item{batch_correction}{Numeric matrix following batch correction with features as rownames and sample names as the column names}

\item{meta}{Data frame of sample data with the first column being sample names that match the column names of the matrix}

\item{batch.1}{Column name from the meta file of the column that will be used for batch information}

\item{variable_of_interest}{Column name from the meta file of the column that will be used for the variable of interest information}

\item{color_by}{Used by the RLE plot to select whether "batch" or "variable_of_interest" will be used to color the individuals. Default is set to "batch"}
}
\value{
A list object of uncorrected and batch corrected RLE plots
}
\description{
Evaluation Relative Log Expression
}
\examples{
set.seed(333)
}
