% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluation_rle.R
\name{evaluation_rle}
\alias{evaluation_rle}
\title{Evaluation Relative Log Expression}
\usage{
evaluation_rle(mat, meta, batch.1, variable_of_interest, color_by, plot_title)
}
\arguments{
\item{mat}{A Numeric matrix or list of matrices after pre-processing and/or batch correction with features as rownames and sample names as the column names}

\item{meta}{Data frame of sample data with the first column being sample names that match the column names of the matrix}

\item{batch.1}{Column name from the meta file of the column that will be used for batch information}

\item{variable_of_interest}{Column name from the meta file of the column that will be used for the variable of interest information}

\item{color_by}{Used by evaluation multiple components, pca, and rle to select which feature will be used to color the individuals. Choices are "batch", "variable_of_interest", "BnW", and "all". Default is set to "all"}

\item{plot_title}{A list of titles for each matrix submitted. Used to named the plots generated by the batch evaluate function. Default will generate titles based on the uncorrected and corrected matrices.}
}
\value{
A list object of RLE plots
}
\description{
Evaluation Relative Log Expression
}
\examples{
set.seed(333)
}
