% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{hsaio}
\alias{hsaio}
\title{Eisenberg's housekeeping genes}
\format{
\subsection{\code{hsaio}}{

a vector with house keeping genes
}
}
\source{
\url{https://www.gsea-msigdb.org/gsea/msigdb/cards/HSIAO_HOUSEKEEPING_GENES}
}
\usage{
hsaio
}
\description{
Eisenberg's housekeeping genes
}
\keyword{datasets}
