% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data}
\alias{merge_data}
\title{merge_data}
\usage{
merge_data(
  merge_matrix_files = NULL,
  merge_meta_files = NULL,
  keep_all_genes = FALSE
)
}
\arguments{
\item{merge_matrix_files}{A list of matrix files to be merged}

\item{merge_meta_files}{A list of meta files to be merged}

\item{keep_all_genes}{A TRUE/FALSE statement to indicate how unmatched genes will be handled. If TRUE, then NAs will be inserted for unmatched genes. If FALSE, unmatched genes will be deleted. Default is set to FALSE}
}
\value{
A list of merged matrix and meta files
}
\description{
merge_data
}
\examples{
set.seed(333)
}
