% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/merge_data.R
\name{merge_data}
\alias{merge_data}
\title{Merge data}
\usage{
merge_data(
  merge_matrix_files = NULL,
  merge_meta_files = NULL,
  keep_all_genes = FALSE
)
}
\arguments{
\item{merge_matrix_files}{A list of matrix files (as data frames) to be merged. If supplied as a named list, batchflex will add a named column to the meta file with the specified names. If not, batch flex will add a generic name per meta file.}

\item{merge_meta_files}{A list of meta files to be merged}

\item{keep_all_genes}{A TRUE/FALSE statement to indicate how unmatched genes will be handled. If TRUE, then NAs will be inserted for unmatched genes. If FALSE, unmatched genes will be deleted. Default is set to FALSE}
}
\value{
A list of merged matrix and meta files
}
\description{
Merges multiple expression matrices and generates a simple meta file to be used by \code{Batch_FLEX}, if none is provided.
If meta files are provided, they will be merged and a simple meta file will not be generated.
Samples from the merged meta file and matrix file will be matched using \code{RecordLinkage::levenshteinSim} generated scores and will be ordered similarly.
}
\examples{
set.seed(333)
matrices_to_be_merged <- list("GSE112876" = BatchFLEX::GSE112876_matrix,
"GSE15907" = BatchFLEX::GSE15907_matrix,
"GSE37448" = BatchFLEX::GSE37448_matrix,
"GSE60336" = BatchFLEX::GSE60336_matrix,
"GSE75195" = BatchFLEX::GSE75195_matrix,
"GSE75202" = BatchFLEX::GSE75202_matrix,
"GSE75203" = BatchFLEX::GSE75203_matrix)
meta_files_to_be_merged <- list(BatchFLEX::GSE112876_meta,
                               BatchFLEX::GSE15907_meta,
                               BatchFLEX::GSE37448_meta,
                               BatchFLEX::GSE60336_meta,
                               BatchFLEX::GSE75195_meta,
                               BatchFLEX::GSE75202_meta,
                               BatchFLEX::GSE75203_meta)
test_merge <- merge_data(merge_matrix_files = matrices_to_be_merged, merge_meta_files = meta_files_to_be_merged)
head(as.data.frame(test_merge$merged_matrix), n = c(5,5))
head(test_merge$merged_meta, n = c(5,5))

}
